Move google analytics to app
