+++ src/google_analytics/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'google_analytics.apps.GoogleAnalyticsConfig'
